!XB (Reflection) Fixed Win32 build. Review: ID 69083 by Author: Dominique Mader, Reviewer: Luis Lairet, Observer: Sebastien Laurent, Observer: Nico Moss  (Approved by luis)
Copied from Perforce
 Change: 1600196

+++ Code/CryEngine/CryCommon/CryReflection/IReflection.h
@@ -128,8 +128,8 @@ private:
 	ValueType m_value;
 };
 
-typedef Index<uint16, ~0> TypeIndex;
-typedef Index<uint16, ~0> ExtensionIndex;
+typedef Index<size_t, ~0> TypeIndex;
+typedef Index<size_t, ~0> ExtensionIndex;
 
 struct SSourceFileInfo
 {
